requirements.txt
tensorboard==2.9.1
pypinyin
pytorch-lightning
scikit-learn
scipy
timm
tokenizers
torch-tb-profiler
torchmetrics==0.10.3
torchvision
tqdm
transformers
urllib3
pyarrow
pydantic
lightning-bolts
rich
deepspeed
sqlitedict